app/src/main/java/com/example/geoffrey/bibliotheekapp/database/BookDatabase.kt
package com.example.geoffrey.bibliotheekapp.database

import android.content.Context
import androidx.room.Database
import androidx.room.Room
import androidx.room.RoomDatabase
import com.example.geoffrey.bibliotheekapp.models.Book


@Database(entities = [Book::class], version = 1, exportSchema = false)
abstract class BookDatabase: RoomDatabase() {
   abstract val bookDatabaseDao:BookDatabaseDao
    companion object {
        @Volatile
        private var INSTANCE: BookDatabase? = null

        fun getInstance(context: Context):BookDatabase{
            synchronized(this) {
                var instance = INSTANCE
                if(instance == null) {
                    instance= Room.databaseBuilder(
                        context.applicationContext,
                        BookDatabase::class.java,
                        "book_database"
                    ).fallbackToDestructiveMigration().build()
                    INSTANCE = instance
                }
                return instance
            }
        }
    }
}